+++ packages/core/src/families/hedera/validation.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+import { schemaFamilies, schemaTransactionCommon } from "../common";
+
+export const schemaRawHederaTransaction = schemaTransactionCommon.extend({
+  family: z.literal(schemaFamilies.enum.hedera),
+  memo: z.string().max(100).optional(),
+});