+++ packages/core/src/families/near/validation.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+import { schemaFamilies, schemaTransactionCommon } from "../common";
+
+export const schemaRawNearTransaction = schemaTransactionCommon.extend({
+  family: z.literal(schemaFamilies.enum.near),
+  mode: z.string(),
+  fees: z.string().optional(),
+});