+++ packages/core/src/families/types.ts
@@ -3,15 +3,23 @@ import type { z } from "zod";
 import type { AlgorandTransaction } from "./algorand/types";
 import type { BitcoinTransaction } from "./bitcoin/types";
 import type { schemaFamilies } from "./common";
+import type { CeloTransaction } from "./celo/types";
 import type { CosmosTransaction } from "./cosmos/types";
 import type { CryptoOrgTransaction } from "./crypto_org/types";
 import type { EthereumTransaction } from "./ethereum/types";
+import type { HederaTransaction } from "./hedera/types";
+import type { FilecoinTransaction } from "./filecoin/types";
+import type { NearTransaction } from "./near/types";
+import type { NeoTransaction } from "./neo/types";
 import type { PolkadotTransaction } from "./polkadot/types";
 import type { RippleTransaction } from "./ripple/types";
+import type { SolanaTransaction } from "./solana/types";
 import type { StellarTransaction } from "./stellar/types";
 import type { TezosTransaction } from "./tezos/types";
 import type { TronTransaction } from "./tron/types";
+import type { ElrondTransaction } from "./elrond/types";
 import type { schemaRawTransaction } from "./validation";
+import type { CardanoTransaction } from "./cardano/types";
 
 /**
  * Supported coin families
@@ -58,9 +66,17 @@ export type Transaction =
   | BitcoinTransaction
   | AlgorandTransaction
   | CryptoOrgTransaction
+  | HederaTransaction
+  | FilecoinTransaction
   | RippleTransaction
+  | CeloTransaction
   | CosmosTransaction
   | TezosTransaction
   | PolkadotTransaction
   | StellarTransaction
-  | TronTransaction;
+  | TronTransaction
+  | NearTransaction
+  | NeoTransaction
+  | ElrondTransaction
+  | CardanoTransaction
+  | SolanaTransaction;